Chore(landing-page): update source path of file
